refs #5097 added proer type for lockID
<==== Mercurial data ====>
Branch: dev_2.4

+++ common/src/utils/thread/mutex_impl.h
@@ -35,7 +35,7 @@ public:
     void afterMutexLocked(
         const char* sourceFile,
         int sourceLine,
-        int lockID );
+        size_t lockID );
     void beforeMutexUnlocked();
 };
