+++ common/src/utils/thread/mutex_lock_analyzer.cpp
@@ -37,7 +37,7 @@ MutexLockKey::MutexLockKey(
     const char* _sourceFile,
     int _sourceLine,
     QnMutex* _mutexPtr,
-    int _lockID,
+    size_t _lockID,
     std::uintptr_t _threadHoldingMutex )
 :
     sourceFile( _sourceFile ),
